Bugfix in PlayStreak
Bug: the first game of a new streak is not counted in the streak,
neither in `games played in a row` nor in `max time spent playing`. Bug
is not present if the game is the first one ever under that time
control.
Step to reproduce: play one rated game using a time control previously
used (e.g. blitz) even though not in the last hour, than go to
`lichess.org/@/<user>/perf/blitz` and note
that both `current streak` at the bottom of the page are still zero.
From the second game the `current streak` of `games played in a row` is
systematically one less than expected, while the other one does not
count the elapsed time of the first game.

Troubleshooting: the bug is due to the call at `Streak.init`
(i.e. `(0, none, none)`) when a game is discontinued from the previous one.
I propose, with this commit, to substitute that call with a new function
inspired by `inc`. This new function is called `res` (from `reset`):
its aim is to set `nb` and `time` dynamically after the first game
of the new streak.